+++ extent/deployment.go
@@ -1,4 +1,4 @@
-package vo
+package extent
 
 type (
 	RemoteDeployment struct {
