+++ extent/webhook.go
@@ -1,4 +1,4 @@
-package vo
+package extent
 
 type (
 	WebhookConfig struct {
